refactor: remove on-idle functionality
Remove the legacy on-idle LLM analysis feature that was disabled by default. This simplifies the codebase by removing:
- on-idle.ts strategy and prompt
- OnIdle config interface and validation
- Event handler for session.status idle events
- Related documentation

The discard/extract tools remain as the primary pruning mechanism.
