+++ index.ts
@@ -1,10 +1,10 @@
 import type { Plugin } from "@opencode-ai/plugin"
 import { getConfig } from "./lib/config"
 import { Logger } from "./lib/logger"
-import { loadPrompt } from "./lib/prompt"
+import { loadPrompt } from "./lib/prompts"
 import { createSessionState } from "./lib/state"
 import { createDiscardTool, createExtractTool } from "./lib/strategies"
-import { createChatMessageTransformHandler, createEventHandler } from "./lib/hooks"
+import { createChatMessageTransformHandler } from "./lib/hooks"
 
 const plugin: Plugin = (async (ctx) => {
     const config = getConfig(ctx)
@@ -91,7 +91,6 @@ const plugin: Plugin = (async (ctx) => {
                 )
             }
         },
-        event: createEventHandler(ctx.client, config, state, logger, ctx.directory),
     }
 }) satisfies Plugin
