+++ lib/hooks.ts
@@ -5,7 +5,6 @@ import { syncToolCache } from "./state/tool-cache"
 import { deduplicate, supersedeWrites, purgeErrors } from "./strategies"
 import { prune, insertPruneToolContext } from "./messages"
 import { checkSession } from "./state"
-import { runOnIdle } from "./strategies/on-idle"
 
 export function createChatMessageTransformHandler(
     client: any,
@@ -32,36 +31,6 @@ export function createChatMessageTransformHandler(
 
         if (state.sessionId) {
             await logger.saveContext(state.sessionId, output.messages)
-        }
-    }
-}
-
-export function createEventHandler(
-    client: any,
-    config: PluginConfig,
-    state: SessionState,
-    logger: Logger,
-    workingDirectory?: string,
-) {
-    return async ({ event }: { event: any }) => {
-        if (state.sessionId === null || state.isSubAgent) {
-            return
-        }
-
-        if (event.type === "session.status" && event.properties.status.type === "idle") {
-            if (!config.strategies.onIdle.enabled) {
-                return
-            }
-            if (state.lastToolPrune) {
-                logger.info("Skipping OnIdle pruning - last tool was prune")
-                return
-            }
-
-            try {
-                await runOnIdle(client, state, logger, config, workingDirectory)
-            } catch (err: any) {
-                logger.error("OnIdle pruning failed", { error: err.message })
-            }
         }
     }
 }
