+++ lib/strategies/index.ts
@@ -1,5 +1,4 @@
 export { deduplicate } from "./deduplication"
-export { runOnIdle } from "./on-idle"
 export { createDiscardTool, createExtractTool } from "./tools"
 export { supersedeWrites } from "./supersede-writes"
 export { purgeErrors } from "./purge-errors"
